requirements.txt
discord.py
python-dot